pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yk-case-generation"
version = "0.0.1"
description = "CLI tool for case generation pipeline"
authors = [{name="YK Team"}]
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn",
  "httpx",
  "pydantic",
  "python-docx",
  "pillow",
  "pdf2image",
  "tencentcloud-sdk-python",
  "pydantic-settings",
  "jsonschema",
  "typer",
  "structlog",
  "pypdf",
  "tenacity",
  "pandas"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-mock",
  "black",
  "ruff",
  "mypy"
]

[project.scripts]
ykcg = "yk_case_generation.cli.__main__:app"

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = 3.11
ignore_missing_imports = true
